requirements.txt
# Flask and extensions
Flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4

# Database
SQLAlchemy==2.0.9
alembic==1.10.3

# Utilities
python-dotenv==1.0.0
Werkzeug==2.2.3
requests==2.28.2
pyjwt==2.6.0
tqdm>=4.65.0
networkx>=3.0

# AI and RAG
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.16,<0.1.0
langchain-core>=0.1.18,<0.2.0
langchain-chroma>=0.0.10
langchain-openai>=0.0.2
langsmith>=0.0.83,<0.1.0
openai==1.3.0
chromadb>=0.4.18
faiss-cpu==1.7.4
tiktoken==0.4.0
sentence-transformers>=2.2.2

# Data processing
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
pymupdf>=1.22.5
unstructured>=0.10.30

# Testing
pytest==7.3.1
pytest-flask==1.2.0 